requirements.txt
fastapi
uvicorn[standard]
pyyaml
pydantic

